Adding tasks is now possible
Created and event that listens for the key pressed on the input field,
in case of it being the enter key the new task will be added to the
list.
The “add to list” button is now calling the same function as the enter
key.

Did some minor changes to styles as well for every angular event
element to have the pointing cursor and expanded the input field a
little bit on larger screens.