Animations
CSS 3 animations included as well as the ngAnimate module to animate
the new tasks.

Unfortunately Angular has some kind of issue with animating the right
element on lists that use orderBy, so I couldn’t animate the entering
tasks on the completed list, since those are ordered by the
completedDate property.