Animation tweaks
Turns out the issue with ngAnimate not animating the right element had
something to do with tracking the loop by the $index.
Added a unique id to each task to solve that.
Added new animations to completing and removing a task, as well as when
it enters the completed list.